Test (QuartzNet): add xfail detection to Windows w/ Numpy 1.18+ (under Nox) for QuartzNet import
